moved usermanagement to DIR
