+++ samjTBC/samj/viewClasses/usermanagement_api.py
@@ -1,0 +1,138 @@
+
+import json
+from django.http import JsonResponse
+from django.views import View
+
+from ..models import GlobalSettings, User
+from ..views import validate_password
+
+from django.utils.decorators import method_decorator
+from django.views.decorators.csrf import csrf_exempt
+
+@method_decorator(csrf_exempt, name='dispatch')
+class UserManagementAPIView(View):
+    def get(self, request, *args, **kwargs):
+        users = User.objects.all().values('username', 'fullname', 'status', 'role', 'number')
+        user_list = list(users)
+        status_options = [choice for choice in User.STATUS_CHOICES if choice[0] != 'deleted']
+        role_options = User.ROLE_CHOICES
+        data = {
+        'users': user_list,
+        'status_options': status_options,
+        'role_options': role_options,
+        }
+        return JsonResponse(data)
+
+    def post(self, request, *args, **kwargs):
+        try:
+            data = json.loads(request.body)
+            password = data.get('password')
+
+            # Validate password if provided
+            if password:
+                is_valid, message = validate_password(password)
+                if not is_valid:
+                    return JsonResponse({'error': message}, status=400)
+
+            # Create default global settings for the new user
+            global_settings = GlobalSettings.objects.create()
+
+            user = User(
+                username=data.get('username'),
+                fullname=data.get('fullname'),
+                password=password,
+                number=data.get('number'),
+                status=data.get('status', 'active'),
+                role=data.get('role', 'user'),
+                global_settings=global_settings
+            )
+            user.save()
+            return JsonResponse({'message': 'User created successfully'}, status=201)
+        except json.JSONDecodeError:
+            return JsonResponse({'error': 'Invalid JSON'}, status=400)
+        except Exception as e:
+            return JsonResponse({'error': str(e)}, status=400)
+
+    def put(self, request, *args, **kwargs):
+        try:
+            data = json.loads(request.body)
+            username = data.get('username')
+            if not username:
+                return JsonResponse({'error': 'Username is required for updating a user'}, status=400)
+            try:
+                user = User.objects.get(username=username)
+            except User.DoesNotExist:
+                return JsonResponse({'error': 'User not found'}, status=404)
+
+            password = data.get('password')
+
+            # Validate password if provided
+            if password:
+                is_valid, message = validate_password(password)
+                if not is_valid:
+                    return JsonResponse({'error': message}, status=400)
+
+            user.fullname = data.get('fullname', user.fullname)
+            if password:
+                user.password = password
+            user.number = data.get('number', user.number)
+            user.status = data.get('status', user.status)
+            user.role = data.get('role', user.role)
+            user.save()
+            return JsonResponse({'message': 'User updated successfully'}, status=200)
+        except json.JSONDecodeError:
+            return JsonResponse({'error': 'Invalid JSON'}, status=400)
+        except Exception as e:
+            return JsonResponse({'error': str(e)}, status=400)
+
+    def patch(self, request, *args, **kwargs):
+        try:
+            data = json.loads(request.body)
+            username = data.get('username')
+            if not username:
+                return JsonResponse({'error': 'Username is required for updating a user'}, status=400)
+            try:
+                user = User.objects.get(username=username)
+            except User.DoesNotExist:
+                return JsonResponse({'error': 'User not found'}, status=404)
+
+            # Fields that may be updated
+            updatable_fields = ['fullname', 'password', 'number', 'status', 'role']
+
+            password = data.get('password')
+            # Validate and update password if provided
+            if password:
+                is_valid, message = validate_password(password)
+                if not is_valid:
+                    return JsonResponse({'error': message}, status=400)
+                user.password = password
+
+            # Update other fields
+            for field in updatable_fields:
+                if field in data:
+                    setattr(user, field, data[field])
+
+            user.save()
+            return JsonResponse({'message': 'User updated successfully'}, status=200)
+        except json.JSONDecodeError:
+            return JsonResponse({'error': 'Invalid JSON'}, status=400)
+        except Exception as e:
+            return JsonResponse({'error': str(e)}, status=400)
+
+    def delete(self, request, *args, **kwargs):
+        try:
+            data = json.loads(request.body)
+            username = data.get('username')
+            if not username:
+                return JsonResponse({'error': 'Username is required for deleting a user'}, status=400)
+            try:
+                user = User.objects.get(username=username)
+            except User.DoesNotExist:
+                return JsonResponse({'error': 'User not found'}, status=404)
+
+            user.delete()
+            return JsonResponse({'message': 'User deleted successfully'}, status=204)
+        except json.JSONDecodeError:
+            return JsonResponse({'error': 'Invalid JSON'}, status=400)
+        except Exception as e:
+            return JsonResponse({'error': str(e)}, status=400)
